fix: database initialization via IAssemblyFixture
