+++ tests/Ticketing.WebAPI.IntegrationTests/Ticketing.WebAPI.IntegrationTests.csproj
@@ -21,6 +21,7 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
+	  <PackageReference Include="xunit.assemblyfixture" Version="2.2.0" />
   </ItemGroup>
 
   <ItemGroup>
